パスワードリセット機能の実装

+++ db/migrate/20241112145728_modify_trees_table.rb
@@ -1,9 +1,16 @@
 class ModifyTreesTable < ActiveRecord::Migration[7.1]
   def change
     # 不要なカラムを削除
-    remove_column :trees, :growth_stage, :integer
-    remove_column :trees, :last_watered_at, :datetime
-    remove_column :trees, :water_count, :integer
+    if column_exists?(:trees, :growth_stage)
+      remove_column :trees, :growth_stage, :integer
+    end
+    if column_exists?(:trees, :last_watered_at)
+      remove_column :trees, :last_watered_at, :datetime
+    end
+    if column_exists?(:trees, :water_count)
+      remove_column :trees, :water_count, :integer
+    end
+    
 
     # 新しいカラムを追加
     add_column :trees, :name, :string
